requirements.txt
Flask==2.0.3
Pillow==9.0.1
requests==2.27.1
torchvision==0.11.3
